+++ lib/commands/commands_list.dart
@@ -27,6 +27,7 @@ final List<Command> commands = [
   InstallCommand(),
   RemoveCommand(),
   SortCommand(),
+  BuildCommand(),
   UpdateCommand(),
 ];