+++ lib/commands/impl/init/flutter/init_getxpattern.dart
@@ -1,7 +1,6 @@
 import 'dart:io';
 
 import '../../../../common/utils/logger/log_utils.dart';
-import '../../../../common/utils/pubspec/pubspec_utils.dart';
 import '../../../../core/internationalization.dart';
 import '../../../../core/locales.g.dart';
 import '../../../../core/structure.dart';