+++ lib/samples/impl/get_route.dart
@@ -5,7 +5,7 @@ class RouteSample extends Sample {
   RouteSample({String path = 'lib/app/routes/app_routes.dart'}) : super(path);
   @override
   String get content => '''part of 'app_pages.dart';
-// DO NOT EDIT. This is code generated via package:get_cli/rgb_cli.dart
+// DO NOT EDIT. This is code generated via package:rgb_cli/rgb_cli.dart
 
 abstract class Routes{
   Routes._();