+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## [1.0.3]
+
+- Update: add help examples
+- 
+## [1.0.2]
+
+- Fix: add missing dependency in pubspec.yaml
+- Fix: add missing import in lib/rgb.dart
+
+## [1.0.1]
+
+- Fix: version check
+- Fix: dev version check
+- Fix: version update
+- Feature: add build command
+- Improvement: set static const variable when create or update routes
+
+
 ## [1.0.0]
 
 - First release